Quote special characters in pillar.example
The way in which both illegal_object_name_chars and
illegal_macro_output_chars were defined in the example pillar caused
rendering errors in Yaml as `:` was interpreted as being part of a
mapping definition. This patch quotes special characters in there
explicitly and thereby happifies the parser.

See #26 for further details.